+++ internal/domain/recurrentmovement/repository/recurrent_movement.go
@@ -3,6 +3,7 @@ package repository
 import (
 	"context"
 	"gorm.io/gorm"
+	"time"
 
 	"github.com/google/uuid"
 
@@ -11,6 +12,7 @@ import (
 
 type RecurrentRepository interface {
 	AddConsistent(ctx context.Context, tx *gorm.DB, recurrent model.RecurrentMovement) (model.RecurrentMovement, error)
+	FindByMonth(ctx context.Context, initialDate time.Time) ([]model.RecurrentMovement, error)
 }
 
 type recurrentRepository struct {
@@ -32,4 +34,36 @@ func (r *recurrentRepository) AddConsistent(ctx context.Context, tx *gorm.DB, re
 		return model.RecurrentMovement{}, err
 	}
 	return recurrent, nil
+}
+
+func (r *recurrentRepository) FindByMonth(ctx context.Context, date time.Time) ([]model.RecurrentMovement, error) {
+	userID := ctx.Value("user_id").(string)
+	var recurrents []model.RecurrentMovement
+
+	err := r.gorm.
+		Select([]string{
+			"recurrent_movements.id",
+			"recurrent_movements.description",
+			"recurrent_movements.initial_date",
+			"recurrent_movements.amount",
+			`w.id as "Wallet__id"`,
+			`w.description as "Wallet__description"`,
+			`c.id as "Category__id"`,
+			`c.description as "Category__description"`,
+			`sc.id as "SubCategory__id"`,
+			`sc.description as "SubCategory__description"`,
+		}).
+		Order("recurrent_movements.initial_date desc").
+		Joins("left join wallets w on recurrent_movements.wallet_id = w.id").
+		Joins("left join categories c on recurrent_movements.category_id = c.id").
+		Joins("left join sub_categories sc on recurrent_movements.sub_category_id = sc.id").
+		Where("recurrent_movements.user_id = ?", userID).
+		Where("recurrent_movements.initial_date <= ?", date).
+		Where("recurrent_movements.end_date >= ?", date).
+		Or("recurrent_movements.end_date is null").
+		Find(&recurrents).Error
+	if err != nil {
+		return nil, err
+	}
+	return recurrents, nil
 }